Move playlists and favorites test files out of subdirectories
