+++ tests/favorites.spec.js
@@ -1,9 +1,9 @@
 var shell = require('shelljs');
 var request = require("supertest");
-var app = require('../../app');
+var app = require('../app');
 
 const environment = process.env.NODE_ENV || 'test';
-const configuration = require('../../knexfile')[environment];
+const configuration = require('../knexfile')[environment];
 const database = require('knex')(configuration);
 
 describe('Favorites endpoints', () => {
